Add logic for looking for available versions of vhs

+++ rec/rec.sh
@@ -5,11 +5,22 @@ cd "$(dirname "$0")"
 
 export VHS_NO_SANDBOX=1
 
+VHS=$(which vhs)
+if [ -x "$HOME/clones/vhs/build/vhs" ]; then
+  VHS="$HOME/clones/vhs/build/vhs"
+elif [ -x "$HOME/clones/vhs/vhs" ]; then
+  VHS="$HOME/clones/vhs/vhs"
+fi
+if [ ! -x $VHS ]; then
+  echo 'Could not find vhs'
+  exit 1
+fi
+
 rm -rf ./tmp
 mkdir -p tmp
 for lang in c cpp rust zig python bash odin go haskell; do
   # github.com/xyproto/vhs has --keypress-overlay support
-  $HOME/clones/vhs/build/vhs simple_$lang.tape --keypress-overlay || vhs simple_$lang.tape
+  "$VHS" debug_$lang.tape --keypress-overlay || "$VHS" debug_$lang.tape
   mv -v simple_$lang.gif gif/
   rm -rf ./tmp
   mkdir -p tmp
